requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==4.0.6
gunicorn==20.0.4
django-crispy-forms==2.2
djangorestframework==3.15.2
pillow==10.4.0
sqlparse==0.5.0